+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.43.0
+### Breaking
+* echo of undefined or Promise that resolve to undefined no longer display string undefined
+### Features
+* add support for `--rows` CSS custom property [#956](https://github.com/jcubic/jquery.terminal/issues/956)
+* add aborting signals [#940](https://github.com/jcubic/jquery.terminal/issues/940)
+### Bugfix
+* fix CMD update performace issue [#961](https://github.com/jcubic/jquery.terminal/issues/961)
+* don't reflow the reflow the cursor on update [#932](https://github.com/jcubic/jquery.terminal/issues/932)
+* fix unexpected uncaught exceptions in promises
+* fix skip/skip_stop return value and add types
+* fix recursive call to echo in renderHandler [#733](https://github.com/jcubic/jquery.terminal/issues/733)
+
 ## 2.42.2
 ### Bugfix
 * actual fix for blank lines typing animation [#946](https://github.com/jcubic/jquery.terminal/issues/946)